consolidated assets and memcache middleware into "cache" middleware
